+++ CHANGELOG.md
@@ -6,6 +6,10 @@ This project follows [pub-flavored semantic versioning][pub-semver]. ([more][pub
 [pub-semver]: https://www.dartlang.org/tools/pub/versioning.html#semantic-versions
 [pub-semver-readme]: https://pub.dartlang.org/packages/pub_semver
 
+## HEAD
+
+- [feat] add optional parameter `quiet` to `DotEnv` constructor
+
 ## 4.1.0
 
 - [feat] add `DotEnv.getOrElse`